Trim the project to keep only the audit client
The project is moving towards a library instead of an agent replacement
for auditd for similar agents. This commit cleans out a lot of code that
was for initializing this as an agent and having examples, samples etc.
It will just need client APIs to be able to read the logs and dump it to
output mechanism instead of a stream like file or syslog.

Right now it will retain writing to stdout since that will be necessary
for inspection